+++ src/main/java/samuelmovi/familyLibraryJava/repo/LocationRepository.java
@@ -5,9 +5,11 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
+import java.util.List;
 
 @Repository
 public interface LocationRepository extends CrudRepository<Location, Long>{
 
+    List<Location> findAll();
     Location findByAddress(String address);
 }